+++ app.rb
@@ -1,99 +1,159 @@
-require './person'
-require './student'
-require './teacher'
-require './book'
-require './capitalize_decorator'
-require './trimmer_decorator'
-require './classroom'
+require './valid_date'
+require './library'
+require './menu'
 
 class App
   def initialize
-    @people = []
-    @books = []
-    @rentals = []
-    @classroom = Classroom.new('101')
+    @library = Library.new
+    @menu = Menu.new
   end
-
-  attr_reader :rentals, :people, :books
 
   def select_book_from_list
     puts 'Select a book from the following list by number'
-    @books.each_with_index do |book, index|
+    @library.books.each_with_index do |book, index|
       puts "#{index + 1}) Title: \"#{book.title}\" Author: #{book.author} "
     end
-    gets.chomp.to_i
+    gets.chomp.to_i - 1
   end
 
   def select_person_from_list
     puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
+    @library.people.each_with_index do |person, index|
       print "#{index + 1}) #{person.is_a?(Teacher) ? '[Teacher]' : '[Student]'} "
       puts "Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
-    gets.chomp.to_i
+    gets.chomp.to_i - 1
   end
 
   def can_create_rental?
-    if @books.empty?
+    if @library.books.empty?
       puts 'There are no books for rentals'
       return false
     end
-    if @people.empty?
+    if @library.people.empty?
       puts 'There are no people for rentals'
       return false
     end
     true
   end
 
-  def create_rental(date, index_person, index_book)
-    @rentals.push(Rental.new(date, @books[index_book], @people[index_person]))
-  end
-
   def list_rentals
-    if @rentals.empty?
+    if @library.rentals.empty?
       puts 'No Rentals to show'
       return
     end
     print 'ID of person: '
     id = gets.chomp
     puts 'Rentals'
-    @rentals.each do |rental|
+    @library.rentals.each do |rental|
       if rental.person.id == id.to_i
-        print("Date: #{rental.date} ")
-        puts("Book \"#{rental.book.title}\" by #{rental.book.author} ")
+        print "Date: #{rental.date} "
+        puts "Book \"#{rental.book.title}\" by #{rental.book.author} "
       end
     end
   end
 
-  def create_book(title, author)
-    @books.push(Book.new(title, author))
-  end
-
   def list_books
-    if @books.empty?
+    if @library.books.empty?
       puts 'There are no books to show'
       return
     end
-    @books.each { |book| puts "Title: \"#{book.title}\", Author: #{book.author}" }
+    @library.books.each { |book| puts "Title: \"#{book.title}\", Author: #{book.author}" }
   end
 
   def list_people
-    if @people.empty?
+    if @library.people.empty?
       puts 'There are no people to show'
       return
     end
-    @people.each do |person|
+    @library.people.each do |person|
       print '[Teacher] ' if person.is_a?(Teacher)
       print '[Student] ' if person.is_a?(Student)
       puts "Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
   end
 
-  def add_teacher(age, name, specialization)
-    @people.push(Teacher.new(age, specialization, name))
+  def create_rental
+    return unless can_create_rental?
+
+    index_book = select_book_from_list
+    if index_book.negative? || index_book >= @library.books.length
+      puts 'Invalid selection'
+      return
+    end
+    index_person = select_person_from_list
+    if index_person.negative? || index_person >= @library.people.length
+      puts 'Invalid selection'
+      return
+    end
+    puts 'Date (yyyy/mm/dd):'
+    date = gets.chomp
+    if valid_date(date)
+      @library.add_rental(date, index_person, index_book)
+      puts 'Rental created successfully'
+      return
+    end
+    puts 'Invalid date'
   end
 
-  def add_student(age, name, permission)
-    @people.push(Student.new(age, @classroom, name, parent_permission: permission))
+  def create_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    @library.add_book(title, author)
+    puts 'Book created successfully'
+  end
+
+  def new_person
+    print 'Do you want to create a student (1) or a teacher (2)? [Input the number]:'
+    option = gets.chomp.to_i
+    return if option < 1 && option > 2
+
+    case option
+    when 1
+      age, name, permission = student_info
+      @library.create_student(age, name, parent_permission: permission)
+      puts 'Person created successfully'
+    when 2
+      age, name, specialization = teacher_info
+      @library.create_teacher(age, specialization, name)
+      puts 'Person created successfully'
+    end
+  end
+
+  def student_info
+    print 'Age: '
+    age = gets.chomp
+    print 'Name: '
+    name = gets.chomp
+    print 'Has parent permission? [Y/N]: '
+    permission = gets.chomp
+    [age, name, permission.downcase == 'y']
+  end
+
+  def teacher_info
+    print 'Age: '
+    age = gets.chomp
+    print 'Name: '
+    name = gets.chomp
+    print 'Specialization: '
+    specialization = gets.chomp
+    [age, name, specialization]
+  end
+
+  def run(option)
+    case option
+    when 1 then list_books
+    when 2 then list_people
+    when 3 then new_person
+    when 4 then create_book
+    when 5 then create_rental
+    when 6 then list_rentals
+    end
+  end
+
+  def main_menu
+    @menu.main_menu
   end
 end
